requirements.txt
fastapi
uvicorn
langchain
langchain-community
qdrant-client
opensearch-py
redis
prometheus-client
pydantic-settings
pandas
